requirements.txt
requests==2.31.0
dnspython==2.6.1
python-whois==0.8.0
Pillow==10.2.0
imagehash==4.3.1
